Move ImpactLoader.cs from GitCommands.Statistics to GitExtensions.Plugins.GitImpact
