+++ Plugins/Statistics/GitImpact/FormImpact.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Windows.Forms;
-using GitCommands.Statistics;
 using GitUI;
 using GitUIPluginInterfaces;
 using ResourceManager;
